requirements.txt
# Alpaca trading API client
alpaca-trade-api>=3.0.0

# Load .env files for API keys
python-dotenv>=1.0.0

# Technical analysis indicators (Bollinger Bands, RSI, EMA, ATR, etc.)
pandas-ta>=0.3.14

# Data manipulation — used everywhere for price data
pandas>=2.0.0

# Numerical computing — used by pandas-ta and our calculations
numpy>=1.24.0

# Timezone handling — needed for trading window checks (US/Eastern)
pytz>=2024.1

# Web dashboard — lets you watch the bot trade from your browser
flask>=3.0.0

# HTTP client — used for Alpaca News API calls
requests>=2.31.0

# Google Gemini API — free tier AI-powered deep analysis
# Get a free API key at https://aistudio.google.com/apikey
google-generativeai>=0.8.0

# Yahoo Finance — free historical data for Oslo Børs backtesting
yfinance>=1.2.0
